split vtt parser into its own file
