+++ background/messages/catch_vtt_site_subtitles.ts
@@ -2,35 +2,8 @@ import type { PlasmoMessaging } from "@plasmohq/messaging"
 
 import type { CatchSiteSubtitlesResponse } from "~background/types/CatchSiteSubtitlesResponse"
 import type { CatchVTTSiteSubtitlesRequest } from "~background/types/CatchVTTSiteSubtitlesRequest"
+import parseVTT from "~background/utils/functions/parseVTT"
 import logDev from "~utils/functions/logDev"
-
-const vttRegex =
-    /(\d{2}:\d{2}:\d{2}\.\d{3}) --> (\d{2}:\d{2}:\d{2}\.\d{3})(?: .*?)?\n+([\s\S]*?)(?=\n\n|\n\d{2}:\d{2}:\d{2}\.\d{3} -->|\Z)/g
-
-const parseVTT = (vttText: string): Record<string, string[]> => {
-    if (!vttText.trim()) {
-        logDev("parseVTT received an empty or invalid string!")
-        return {}
-    }
-
-    logDev("Parsing VTT content:\n", vttText) // Log the full text before parsing
-    const matches = [...vttText.matchAll(vttRegex)]
-    const grouping: Record<string, string[]> = {}
-
-    for (const match of matches) {
-        const startTime = match[1]
-        const textContent = match[3]?.trim().replace(/\n/g, " ")
-
-        if (!textContent) continue
-
-        if (grouping[startTime]) {
-            grouping[startTime].push(textContent)
-        } else {
-            grouping[startTime] = [textContent]
-        }
-    }
-    return grouping
-}
 
 const handler: PlasmoMessaging.MessageHandler<
     CatchVTTSiteSubtitlesRequest,
